update: path control now more managed for production
