+++ springBackend/src/main/resources/application.properties
@@ -7,3 +7,4 @@ logging.level.org.springframework.security=TRACE
 
 security.jwt.secret-key=${JWT_SECRET_KEY}
 security.jwt.algorithm=${JWT_ALGORITHM}
+frontend.url=${FRONTEND_URL}
